+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- forward logind's [PreparingForShutdown](https://www.freedesktop.org/wiki/Software/systemd/inhibit/)
+  to `systemctl/hostname/preparing-for-shutdown`
+- log shutdown [inhibitor locks](https://www.freedesktop.org/wiki/Software/systemd/inhibit/)
+
+### Fixed
+- explicit timestamp type specification to avoid
+  `OverflowError: Python int too large to convert to C long`
 
 ## [0.1.1] - 2020-06-18
 ### Fixed